Move AWS under clients/aws
