+++ app/api/api_v1/routers/search.py
@@ -18,6 +18,7 @@ from sqlalchemy.orm import Session
 from starlette.responses import RedirectResponse
 
 from app.api.api_v1.schemas.search import SearchRequestBody, SearchResponse
+from app.clients.aws.aws import S3Client, S3Document, get_s3_client
 from app.clients.db.session import get_db
 from app.config import (
     AWS_REGION,
@@ -29,7 +30,6 @@ from app.config import (
     VESPA_SECRETS_LOCATION,
     VESPA_URL,
 )
-from app.core.aws import S3Client, S3Document, get_s3_client
 from app.core.custom_app import AppTokenFactory
 from app.core.download import create_data_download_zip_archive
 from app.core.search import (
